requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel>=0.29.0

django==1.10.5

# REST API
djangorestframework==3.5.3
Markdown==2.6.7
django-filter==1.0.1

# Configuration
django-environ==0.4.1

# Images
# Pillow==4.0.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Time zones support
pytz==2016.10

django-cors-headers
